refactored Metamask bar and progress bar into notifications bar
