+++ react-app-typescript/src/containers/Faucet/saga.ts
@@ -4,7 +4,7 @@ import { fetchInfo, getSKYMTokens } from '../../lib/faucet-service';
 import { waitTxConfirmed } from '../../lib/web3-service';
 import { actionTypes as at } from './constants';
 import { fetchInfoSuccess, fetchInfoError } from './actions';
-import { progressMessageAction, progressMessageDoneAction } from '../ProgressBar/actions';
+import { progressMessageAction, progressMessageDoneAction } from '../Notifications/actions';
 import { alertSuccessAction, alertErorrAction  } from '../Alert/actions';
 import { FaucetInfo } from '../../lib/model';
