+++ react-app-typescript/src/containers/Notifications/actions.ts
@@ -2,13 +2,13 @@ import { actionTypes as at } from './constants';
 
 export const progressMessageAction = (text: string) => {
     return {
-        type: at.PROGRESS_MESSAGE,
+        type: at.NOTIFICATIONS_PROGRESS_MESSAGE,
         payload: text
     };
 };
 
 export const progressMessageDoneAction = () => {
     return {
-        type: at.PROGRESS_MESSAGE
+        type: at.NOTIFICATIONS_PROGRESS_MESSAGE
     };
 };
