+++ react-app-typescript/src/containers/Notifications/constants.ts
@@ -1,0 +1,5 @@
+/* Action types */
+export const actionTypes = {
+    NOTIFICATIONS_PROGRESS_MESSAGE: 'NOTIFICATIONS/PROGRESS_MESSAGE',
+    NOTIFICATIONS_PROGRESS_MESSAGE_DONE: 'NOTIFICATIONS/PROGRESS_MESSAGE_DONE'
+};
