+++ react-app-typescript/src/containers/Notifications/model.ts
@@ -1,0 +1,4 @@
+export interface NotificationsAction {
+    type: string
+    payload?: any
+}
