+++ react-app-typescript/src/containers/Notifications/reducer.ts
@@ -1,18 +1,18 @@
 import { fromJS } from 'immutable';
 
 import { actionTypes as at } from './constants';
-import { ProgressBarAction } from './model';
+import { NotificationsAction } from './model';
 
 const initialState = fromJS({
     message: undefined
 });
 
-export default (state = initialState, action: ProgressBarAction) => {
+export default (state = initialState, action: NotificationsAction) => {
     switch (action.type) {
-        case at.PROGRESS_MESSAGE:
+        case at.NOTIFICATIONS_PROGRESS_MESSAGE:
             return state
                 .set('message', action.payload);
-        case at.PROGRESS_MESSAGE_DONE:
+        case at.NOTIFICATIONS_PROGRESS_MESSAGE_DONE:
             return initialState;
         default:
             return state;
