--- react-app-typescript/src/containers/ProgressBar/constants.ts
@@ -1,5 +1,0 @@
-/* Action types */
-export const actionTypes = {
-    PROGRESS_MESSAGE: 'PROGRESS/MESSAGE',
-    PROGRESS_MESSAGE_DONE: 'PROGRESS/MESSAGE_DONE'
-};
