--- react-app-typescript/src/containers/ProgressBar/model.ts
@@ -1,4 +1,0 @@
-export interface ProgressBarAction {
-    type: string
-    payload?: any
-}
