+++ react-app-typescript/src/root-reducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux-immutable';
 
 import faucetReducer from './containers/Faucet/reducer';
 import faucetAdminReducer from './containers/FaucetAdmin/reducer';
-import progressBarReducer from './containers/ProgressBar/reducer';
+import notificationsReducer from './containers/Notifications/reducer';
 import alertReducer from './containers/Alert/reducer';
 import listingsReducer from './containers/Listings/reducer';
 import uploadReducer from './containers/Upload/reducer';
@@ -10,7 +10,7 @@ import uploadReducer from './containers/Upload/reducer';
 export interface State {
   faucetDomain: any;
   faucetAdminDomain: any;
-  progressBarDomain: any;
+  notificationsDomain: any;
   alertDomain: any;
   listingsDomain: any;
   uploadDomain: any;
@@ -19,7 +19,7 @@ export interface State {
 export const state = combineReducers({
   faucetDomain: faucetReducer,
   faucetAdminDomain: faucetAdminReducer,
-  progressBarDomain: progressBarReducer,
+  notificationsDomain: notificationsReducer,
   alertDomain: alertReducer,
   listingsDomain: listingsReducer,
   uploadDomain: uploadReducer
